requirements.txt
fastapi[standard]
pyjwt
psycopg[c]
python-dotenv